Use a custom DerivedData directory layout
Before this change there was potential for output file clashes. This happened because Bazel has nice isolation in its output directory, while Xcode's default one is very flat. Instead of trying to rename things to work with Xcode's default layout we are now matching most of Bazel's layout.

This requires setting more search paths as not everything is sitting in the same place anymore.