Copy generated files to Derived Data
Before this we would have to setup two search paths for headers in order to find Swift generated header files. This could lead to compilation errors though, as the paths in modulemaps might refer to the generated headers, leading to two different PCMs being created.

Now we copy all generated files to Derived Data. We also setup some symlinks to enable paths in modulemaps to continue to work and we rewrite modulemaps to point to some of those symlinks. Finally we change the reference for Bazel Generated Files in the Project navigator to use a symlink, as `BUILT_PRODUCTS_DIR` is different for each target and using something like `BUILD_DIR` causes the navigator to make files inaccessible.