Fix generated files target dependency detection (#236)
As we add new inputs, such as we did with the Info.plist, it was easy to forget to update the logic around adding the Bazel Generated Files target dependency. Since we have all of the information we need in `input_files.to_dto()`, I've moved the logic there so we will automatically catch these dependencies in the future.