Fix external link for targets without generated files
We need the external symlink for our search paths to external repo headers to work. Before this change we would only create the symlink as part of the Bazel Generated Files target. Now we have a new target that runs before all other targets that ensures that the symlinks exist.

Ideally this would only run if the symlinks didn't exist, but Xcode doesn't support the Input Files being symlinks, so it always runs. Thankfully it's super fast.