Set `SUPPORTED_PLATFORMS` build setting (#289)
For now this is set to the exact platform that the target is configured for. This locks in simulator or device builds.

This is needed to work around an indexing issue. In the future we will support multiple configurations of the same target and `SUPPORTED_PLATFORMS` can be an array then.