Pre-generate files during initial project creation (#298)
To make initial experiences nicer, we now run the `Bazel Generated Files` scheme as part of initial project creation. This won't be run each time the project is generated, only if the `gen_dir` isn't found.

To disable this behavior, you can set `pre_generate_files = False` on your `xcodeproj` target. In that case, it will only run the `Setup` scheme, which will ensure that Xcode behaves nicely once the generated files are created, without having to close and re-open the project.