Add `bazel` option to `build_mode` and force `dbg` for it
Right now the only difference is that in `bazel` mode the `compilation_mode` is currently forced to be `dbg`. This "restriction" will be lifted when we implement proper configuration support. For now though, we need this compilation mode to enable debugging in Build with Bazel.

We now also generate another set of fixtures for Build with Bazel, a.k.a "bwb".