Copy generated files to `$BUILD_DIR/bazel-out` (#327)
This is the same as `$GEN_DIR` for normal builds, but is different for Index Builds. `PBXBuildFile`s will use the `$GEN_DIR` version, so indexing might get messed up until they are normally generated. It's the best we can do though as we need to use the `gen_dir` symlink.