Use Objective-C instead of Swift for compile stub
This resolves some PCM related warnings, but also makes handling these types of targets easier in Build with Bazel mode.

Note, this change does cause the resource bundles to have `OTHER_LDFLAGS` set, but it does no harm, and for bundles that do have executables it would be needed anyway.