Show error when not using tree artifacts (#389)
Not using tree artifacts is a lot slower, so we want to default to using them. If a project needs to use them for some reason, they can pass `archived_bundles_allowed = True` to `xcodeproj`.

We change //examples/ios_app to use tree artifacts while keeping the others as they were. This will allow us to test both paths in Build with Bazel.