Improve debugging of Bazel built projects
This makes three improvements:

1. Sets the working directory to the exec root instead of the project directory. When using `oso_prefix_is_pwd` the paths to archives start with `bazel-out`, which at the project root relies on the fact that the symlink exists, and isn't named something else. At the exec root `bazel-out` is the actual directory.

2. Sets an `target.source-map` entry for `./bazel-out/` to allow breakpoints to hit for generated code.

3. Changes the `target.source-map` entry for `./external/` to `$BAZEL_EXTERNAL`, ensuring that breakpoints work when set from opening a file in the project navigator.

There are still some breakpoint improvements to be made. In particular, when opening a file via Jump to Definition, it will be the fully resolved path. Setting a breakpoint on these will result in a path that these source-maps won't match. I'm working on a fix for that now.