Compare generated files by inclusion and not equality (#441)
In some cases, it seems like "filelist" can contain multiple files built for different configurations, which makes build fail. For now, we can simply check that all defined "rsynclist" files at least exist in "filelist" and ignore generated files for a different configuration in this check.
We likely need to fix this properly based on the target exclusion logic.