Add experimental `device_and_simulator` rule (#507)
This rule is experimental because it might not be the right API to enable proper device support, but it's better to get something out there so we can iterate on the support.

Also updated the multiplatform example to use the `device_and_simulator` rule. Of note, the `//examples/multiplatform/Tool` target is not included in `//examples/multiplatform:device_targets`, as it would generated two configurations of the same binary (because the `swift_binary` doesn't have the rules_apple transition). It's things like this that make me feel this API isn't quite done yet.