Add experimental rule to specify Xcode Managed provisioning profiles (#523)
BwB doesn't actually use the values that this provides to Xcode, besides for letting Xcode produce provisioning profiles for Automatic Signing. So if a project doesn't use Automatic Signing, and uses BwB, it doesn't need this new rule.

BwX though needs to use `local_provisioning_profile` and/or this rule for Manual Signing, in order to let Xcode know the required values it needs to code sign. And the profile will also have to exist at `~/Library/MobileDevice/Provisioning Profiles`. For Automatic Signing this rule will need to be used (probably with `local_provisioning_profile`).