Fix reinstalls of Built with Bazel apps to device (#531)
If an Info.plist file is set Xcode will do post processing on it after all other build steps. This currently changes the order of the keys, which results in a different code signature, breaking validation and thus re-installs to devices. For now we don't associate the Info.plist to Device platforms when Building with Bazel.