Work towards removing the need for `InputFileAttributesInfo.excluded`
This change causes import frameworks and libraries to be included in `extraFiles`, even though they are already included in `linkerInputs`. This is already the case for modulemaps as well, they are already included in `modulemaps`. In a future change we can probably prune things from `extraFiles` that we have collected elsewhere.

A refactor of how we handle resource file collection will allow us to fully remove `InputFileAttributesInfo.excluded`, and the `_SUSPECT_GENERATED_EXTENSIONS` check as well.