Fix breakpoints set from swiftsourceinfo views
swiftsourceinfo files use the non-Project navigator symlinked paths, so we need to add those versions to lldb's `target.source-map` as well.

This is particularly noticeable when using Focused Projects, as the swiftsourceinfo might be one of the main ways you navigate to a module's source code to set a breakpoint.