Include outputs of non-Xcode projects in the generated inputs output group
This enables a sort of automatic Focused Project Build with Xcode mode, where it builds part of the build tree with Bazel that we can't translate properly into an Xcode target.

This isn't perfect yet though, as some of the dependencies of those targets might get made into Xcode targets from different branches in the build graph. Xcode based indexing also doesn't run for those targets, so the need for importing Bazel's index is higher. But since this unblocks some builds, it's worth getting in now and iterating on it.