Fix `.swiftformat` configuration and add first pass for `swiftlint` (#676)
swiftformat: Do not perform any indenting as it indents multi-line string values. The --indentstrings false only ensures that the string is not indented more than the start line.
swiftlint: First pass at config. It excludes test directories and disables some rules that are not relevant for this repository. There are other warnings and errors. However, they may be legitimate. If we enable a CI check, we will want to address those.