Generate all input files in BwX Index Build
Unlike in BwB mode, we don't get a nice "off-by-one" indicator of which targets Index Build has been attempted on in BwX mode. As a work around we are now building all generated inputs in BwX Index Build. This shouldn't be too bad in a normal project, but focused projects might include a lot of extra targets. Hopefully down the line we can find a better solution for both BwX and BwB.