Limit files generated when using custom schemes
Before this change `BazelDependencies` would generate files for all of the targets in a scheme, even when only a subset of targets were being built. Now we leverage the cached data in `XCBuildData` to gain insight into what targets are actually being built, allowing us to only build what is needed.

Sadly this cached data doesn't say which exact target is being built as part of Index Build, so we still have to rely on our off-by-one logic there. Also of note, SwiftUI Previews build every target in a scheme, and this behavior is preserved with this method.