Fix issues not navigating to source files in BwB mode (#893)
Bazel using relative paths breaks Xcode's handling of error paths in the build log. Also, `external/` and `bazel-out/` prefixes need to be translated to be based on the `$BAZEL_OUT` and `$BAZEL_EXTERNAL` build settings, which are probably rooted somewhere in `/var/tmp`.

As a bonus, it strips out color ANSI codes in the log, and forces color to be on. This brings color back to BES logs.

Thank you @keith for providing me with a baseline implementation of this!