Fix linking BwB SwiftUI Previews
Since we no longer use symlinks for generated input files, the SwiftUI Previews build can fail on Info.plists not existing. The current generated `link.params` points to the normal output base, which won't have the dynamic replacement system. Using a separate output base for SwiftUI Previews was more of an optimization anyway, so it's fine to undo this.