Add last updated dates to design docs
This is to make it clear that these are not living documents, but a snapshot in time when a design was being proposed. They are historical, and useful to know why something was implemented the way that it was. The implementations might/probably will change in the future without the documents being updated.