Ensure expand variables context is set for test actions (#956)
Add expand_variables_based_on to xcode_schemes.test_action(). Allows the client to specify a test target to use as the context for environment variable expansion.
If expand_variables_based_on is not specified, one of the test targets will be selected at random.
If a client does not want a context to be used, the can specify the string value none.
Introduce XcodeScheme.VariableExpansionContext to represent the value of expand_variables_based_on.
Introduce XCSchemeInfo.VariableExpansionContextInfo to represent the fully resolved target for the expansion context.
Add macroExpansion to XCSchemeInfo.TestActionInfo.