Add support for "Compile File" command
This PR changes how the `calculate_output_groups.py` script works, having is use knowledge of the destination platform to select the correct `BAZEL_TARGET_ID`s for all `configuredTargets`, and only using the scheme target ids as a backup.

It also has an unused implementation to pick the library target of a merged target. It's unused though because Xcode doesn't update the `buildRequest` reliably for it, and we can get stuck building the wrong output group.