Make the generator universal and fixed minimum OS
For local development of the generator we use `macos_command_line_application` instead of `apple_universal_binary` + `swift_binary`. This is to keep configurations the same between the generator and the tests, which makes the Xcode development experience better. If we used `swift_binary` or `apple_universal_binary` in `xcodeproj`, then the `macos_unit_test` transition (which is used to be able to set a minimum os version on the tests) will create slightly different configurations for our `swift_library`s. Maybe https://github.com/bazelbuild/bazel/issues/6526 will fix that for us.