Include unfocused dependencies in every output group
Before this change, in BwX mode, if an automatically unfocused target was only part of some dependency trees (e.g. a dep of a unit test, but not the test host), then its dependencies would correctly be marked as a unfocused target's dependency, and be set to be compiled with Bazel, but they wouldn't appear in every output group they needed to be in (e.g. the test host's).

Now if a target is supposed to be built with Bazel because it's an automatically unfocused target's dependency, they will appear in the output group of all dependent targets.