Add support for custom scheme pre and post actions (#1047)
Resolves #1025.

This PR adds Pre and Post actions for the scheme steps. We generally use
this for logging some build analytics/tracking and hence this becomes
important for such use-cases.

- The PR extends the existing build action to also take in the actions
(apart from the targets) as the actions are tied to the Build step of
the scheme and are not a separate entity.
- I have currently tested this for both BwX and BwB modes and things
work as expected.