Use `additionalPaths` to better handle target merging (#1160)
Now that we have the `additionalPaths` technology, we don't have to
change the `packageBinDir` of a top-level target. Instead we add the
paths of the merging target to the merged target's `additionalPaths`,
which cause them to be found with `xcodeGeneratedFiles`.

To ensure that `.swiftmodule` lookup continues to work, we adjust the
assignment to `xcodeGeneratedFiles` to use the merged target's
`packageBinDir`. Also, if it's a framework target, we adjust the path to
be inside of the `Modules` directory of the produced framework.