Add Scheme diagnostics (#1196)
This PR adds support for custom schemes specifying the diagnostics that
the scheme should enable/disable by default.

This was a todo in the previous PRs. 

Currently, I find that the generator is using the custom scheme and the
same can be modified to test this PR:
```
  diagnostics = xcode_schemes.diagnostics(
                    sanitizers = xcode_schemes.sanitizers(
                        thread = True,
                        undefined_behavior = True,
                    ),
                ),
```

**TODO:**
1. Once this is landed, there would be a different PR to add custom
scheme support for `examples/sanitizers`.