Nest output base inside of "outer" output path
This is to make sure that the analysis cache stays as hot as possible. This should bring the consistency of BwB performance more in line with BwX.

We currently don't set a different output base for each of the different configs (e.g. SwiftUI Preview, asan, etc.), as some generated build inputs (e.g. Info.plists) currently can't be found that way. This means that switching between SwiftUI Previews and normal builds, or having builds that have different sanitizers set, will clear at least some of the analysis cache.