Collect linker flags from `Action.argv`
This is a step towards using `Action.args` as the source of truth for all linking. We need to migrate `FilePathResolver` and target merging into Starlark before we can fully migrate. For now though, this exposes linker flags that were previously not detected, because we didn't have access to the final linker context.

Also adds an example of previously ignored linkopt setting.