Apply normal configuration segregation to `BUILD_DIR`
When copying over frameworks for SwiftUI Previews, if a framework is used by both a watchOS app and the iOS app (highly likely) they would both try to copy to the same location. This can result in runtime errors as the wrong version might be copied last. Since we don't copy Bazel generated files anymore, this is safe to do: all of the files generated by Xcode will be in the same configuration, or Xcode will know to use the other configuration (such as when bundling a watchOS app).