Add fewer target dependencies in BwB mode
A know regression this currently introduces is that in BwB mode SwiftUI Previews in non-application extension dependencies (e.g. App Clips, Frameworks, watchOS Apps...) won't work, and you'll need to switch to a scheme that includes them. If your custom scheme already includes them it will work. I'm trying to decide the best way to work around it, and it might be an option that you opt into/out of.