Move where `bazel build` happens for BwB Index Builds
Now that we (generally) don't have target dependencies, we can have `bazel build` happen in each target, for Index Builds. We can only make this change for Index Build, because normal builds might try to build multiple targets, and in that case we still use the scheme to tell use all of the targets being built and use a single `bazel build`. Index Build only tries to build a single target at a time, regardless of how schemes are set up, so we can fix our off-by-one target detection and just perform `bazel build` in the target.