Fix horrible BwX `FilePathResolver` performance
We were hashing `FilePathResolver` as part of `MemoizationKey`, to ensure that if there were multiple versions of `FilePathResolver` then the values wouldn't conflict. That itself isn't much of an issue, except in BwX mode that would include a large `xcodeGeneratedFiles` dictionary. Since there is only one instance of `FilePathResolver`, we don't need to include that in the hash, sidestepping this issue. In the future `FilePathResolver` will be in Starlark, removing all of this anyway...