Fix horrible BwX `FilePathResolver` performance (#1360)
Fixes #1359.

Regressed with 5febe757f6c771e32b82bf640f3db5b2f65d34ab.

We were hashing `FilePathResolver` as part of `MemoizationKey`, to
ensure that if there were multiple versions of `FilePathResolver` then
the values wouldn't conflict. That itself isn't much of an issue, except
in BwX mode that would include a large `xcodeGeneratedFiles` dictionary.
Changing `FilePathResolver` to a class and embedding the memoization
into the instance side-steps this issue. In the future
`FilePathResolver` will be in Starlark, removing all of this anyway...