Remap BwB framework paths regardless of target focusing
Before this change, `PREVIEW_FRAMEWORK_PATHS` would change depending on if the linked framework was a focused target or not. We want to always reference the full framework bundle, to ensure that bundled resources are included. Now we generate this mapping in Starlark before target focusing.

Also add more to the UI framework example to show that resources from other frameworks continue to appear.