Remove `xcodeproj_aspect` dependency on `//xcodeproj/internal:build_mode`
We get around this by creating two versions of the aspect, and two versions of `xcodeproj` rules (two use the different versions of the aspect). I'm choosing to do it this way , because some things are much more optimal being done at the aspect level instead of at the final rule level (e.g. unfocused Xcode targets require additional work in BwX mode that we don't do in BwB mode).