Set `PREVIEW_FRAMEWORK_PATHS` in Starlark
This is progress towards moving all file path resolution into Starlark.

For now this makes the spec.json larger, but that can get reduced in the future by making `copy_outputs.sh` generated by Bazel with the paths directly embedded into it, on demand as part of the `bp` output group.