Move indexstores filelist into target's package
Users have reported that the current indexstore filelist path can grow to be too large. This is because it currently encodes the path to the target, and the file will be at the path of the generator. Now the file will be at the path of the target, and the filename is a fixes size.