Add support for bzlmod (#1502)
This PR adds initial support for bzlmod. I still have to test it with an
example project, but the basic infrastructure should be ready.

bzlmod is only really ready to be used starting in Bazel 6.0. For this
reason, the `MODULE.bazel` depends on rules_apple 2.0.0, which is also
the release that requires Bazel 6. I will need to first check the newest
version of rules_apple and rules_swift into the Bazel Central Registry,
and then we can proceed testing the rules_xcodeproj integration.