Include framework search paths in C/C++ flags (#1583)
I thought in f0c637f6bc12be18e8cddb2c5ec03ee11436b47a that these were
already coming through. So this is a continuation of that commit.

This doesn't use `generated_framework_search_paths`, but currently it
doesn't need it. `generated_framework_search_paths` is needed for
linking, because the linked binary can be at a `$(BUILD_DIR)` rooted
path. For headers though, they will all be in the Bazel execution root.
I don't think this changes with static framework targets, but if so we
will tackle that when we add support for them (#779).