Stop tracking header search paths (#1632)
Now that we fully get these from passed through/processed opts, we no
longer need to track them. This will allow for future removal of logic.

We had to slightly change the logic around creating the
xcode-overlay.yaml. Now if there are any C/C++ sources, it will be
created. This should be a very minimal over-shoot for BwX mode only,
with reduced logic.