Fix BwB UI Test Swift debugging
lldb reports the target triple for the UI test runner, instead of the tests, which can result in mismatches on the os version portion of the triple. The main purpose of using the triple was to get the os and arch matching, so we've removed the os version from our matching logic.