Restrict target merging for non-Swift targets as well
It used to be the case that only swiftmodule search paths were an issue, but linker output paths matter as well, which non-Swift targets have.

I think there is a way to lift this restriction in the future, with further removal of `xcodeGeneratedPaths` from `generator`, but for now this fixes certain project generations.