Use resolved execution root for `-vfsoverlay` (#1726)
Similar to how `-vfsoverlay` doesn't have `-working-directory` applied
to it, it seems that other VSF overlay mappings aren't applied to it.
This causes issues in indexing when a generated VFS overlay file is used
(such as in rules_ios rules). So now we use a new build setting to
ensure the correct version of the execution root is used for
`-vfsoverlay` flags from Bazel.