Improve efficiency of clang opt processing
While profiling Starlark on a large project, `startswith` calls in `_process_swiftcopts` took up a lot of time. Most of those are for clang opts handling, so short circuiting that when possible speeds it up. Also, the order that the opts were checked was adjusted to reflect the likelihood of matching.

Also fixed handling of absolute paths for `-I`, `-iquote`, and `-isystem` clang opts.