Replace `__BAZEL` placeholders in compiler flags in the generator (#1832)
This is an optimization since we can do a single replacement pass for
each build setting. Swift is faster than Starlark here as well. Finally,
in the future when we can parallelize the work we do for each target,
this can be done in parallel instead of all serially by Starlark.