Fix source code viewing in Instruments for `generator` (#1859)
This results in cache misses, since the paths are absolute to the
machine they are run on. Ideally we can come up with a way to handle the
relative paths (like launching
`Instruments.app/Contents/MacOS/Instruments` from the Bazel execution
root), but for now this makes it easier for me to profile things.