Simplify `bazel_path` resolution
We don’t need to filter `$PATH`, because resolving to the downloaded `bazel` is fine, and will only happen if a `tools/bazel` isn’t being used (as that would be at the front of the path instead).

We don’t need to unset `BAZELISK_SKIP_WRAPPER`, because we are either resolving to `tools/bazel`, in which case it has no effect (and we probably actually want it set anyway), or we are resolving directly to bazel (in which case it has no effect), or we are resolving to bazelisk (in which case there wasn’t a wrapper anyway, so it should be set for efficiency sake).

Signed-off-by: Brentley Jones <github@brentleyjones.com>