Remove `FilePath.forceGroupCreation`
We already normalize paths in Starlark, so `filePath.path.isFolderTypeFileSource` is a no-op check.

This does codify (for now) that localized files or CodeData files can never have their folders be folders, they will always be recognized as folder type files. Later, after refactoring `createFilesAndGroups`, we can fix this to allow those to be unstructured files as well.

Signed-off-by: Brentley Jones <github@brentleyjones.com>