Remove unneeded `target.files.to_list()`
With other changes I think this is no longer needed (fixtures have no changes), and this reduces memory usage (even though it’s not retained, it’s adding more memory to the Starlark side it seems). I believe the original reason for having it was something involving `*_import` based rules.

Signed-off-by: Brentley Jones <github@brentleyjones.com>