Make iOS tests in examples manual (#2205)
#2196 has surfaced an issue related to tests actually running as part of
the CI checks. There's timeouts happening and we might want to
investigate that better (and potentially decide if we even want to run
tests or not). From a project generation perspective we don't
necessarily have to run the tests.

For those reasons, this PR makes tests `manual`. Planning to do the same
with the new tests being added in #2196 once this lands.

Signed-off-by: Thiago Cruz <thiago.hmcruz@gmail.com>