Process Swift compiler flags during execution phase (#2208)
This is bringing the handling of Swift compiler flags to be closer to
how we handle CC compiler flags. This is also a step towards allowing
`swift_debug_options.py` to take the params files as inputs instead of a
potentially very large command-line. Finally, this is needed for the
future incremental generators.

---------

Signed-off-by: Brentley Jones <github@brentleyjones.com>