Move generator package directory to /var/tmp
This makes it so the generator package will persist across reboots and for a longer period of time. macOS can clean up periodically contents in `/tmp`. This can cause situations where for example a project is generated and the following day the package directory contents won’t exist anymore causing build failures.

Signed-off-by: Patrick Balestra <me@patrickbalestra.com>