Import indexes in Index Builds as well (#2283)
BwB Index Builds also generate unit files, because we have them use the
same build as a normal build in order to get cache hits. If a project
benefits from index importing (focused projects in particular), then
those generated unit files can be imported. This can be beneficial to
getting indexing bootstrapped quicker, since it doesn't require a normal
build of every top level target.

A downside to importing during Index Build is that it adds blocking time
to the end of those builds. We can't do the normal async importing
because then the followup Index Builds will cancel the import of the
previous ones.

So far people think this additional time isn't a hinderance, so we are
now going to do the additional import.

Signed-off-by: Brentley Jones <github@brentleyjones.com>