Use the module name as the default target name
The name of the product module name should be used as a better source for the Xcode target than the target part of the label. This is because the target part of the label does not necessarily have to be unique; only the complete label serves as the unique identifier of the target.

For this reason, rules_xcodeproj should use the name of the product module for disambiguation and as the Xcode target name if it is both set and unique. If it is not unique or not set, we must fall back on the old logic of using the target name.

Signed-off-by: Vadim Smal <vadims@spotify.com>
Signed-off-by: Brentley Jones <github@brentleyjones.com>