Fix disambiguating of targets with multiple Xcode configurations
A `Target` with multiple Xcode configurations should really be seen as multiple targets that share the same `TargetID`. This change reflects that, fixing a certain disambiguation.

I also used the opportunity to make a better key type, which is more efficient. A future change will prevent having to create this key multiple times.

Signed-off-by: Brentley Jones <github@brentleyjones.com>