Use the module name as the default target name (#2407)
The name of the product module name should be used as a better source
for the Xcode target than the target part of the label. This is because
the target part of the label does not necessarily have to be unique;
only the complete label serves as the unique identifier of the target.

For this reason, rules_xcodeproj should use the name of the product
module for disambiguation and as the Xcode target name if it is both set
and unique. If it is not unique or not set, we must fall back on the old
logic of using the target name.

Here is a before and after for `examples/rules_ios`:

| Before | After |
| - | - |
| <img width="289" alt="Before"
src="https://github.com/MobileNativeFoundation/rules_xcodeproj/assets/158658/a2aabeec-6ee1-4ad7-bf9a-7b3e8e8d1fa7">
| <img width="288" alt="After"
src="https://github.com/MobileNativeFoundation/rules_xcodeproj/assets/158658/790e2c86-f660-4841-9482-9dfb4baaa4a3">
|

Signed-off-by: Vadim Smal <vadims@spotify.com>
Signed-off-by: Brentley Jones <github@brentleyjones.com>
Co-authored-by: Vadim Smal <vadims@spotify.com>