Change how host output groups are set in fallback case
Since the `BAZEL_HOST_` build settings aren’t read in `calculate_output_groups.py`, it’s better to have the fallback values determined in a similar way, by having each buildable contribute to the file. We don't include additionalBuildableReferences because those are transitive dependencies and are only in the scheme for Swift UI Previews support.

Signed-off-by: Brentley Jones <github@brentleyjones.com>