Add `PBXProductType.resourceBundle`
A resource bundle and a executable bundle both use the `com.apple.product-type.bundle` product identifier. In order to support encoding the distinction between the two, without adding an `isResourceBundle` attribute like before, while also taking the opportunity to compress the information into a single character, we now use a different raw value for the `enum` and translate it to an identifier where needed.

Signed-off-by: Brentley Jones <github@brentleyjones.com>