Update CI to test Bazel LTS and HEAD (#2568)
Dev tests run what is in `.bazelversion` (which might become pinned to a
version of 7 in the future). But we’ve changed Bazel 7 tests to be
`last_green` (basically HEAD) instead, and added Bazel LTS (latest) as
well. This is a similar setup to rules_apple and rules_swift, which
should allow us to detect issues sooner with Bazel HEAD changes.

---------

Signed-off-by: Brentley Jones <github@brentleyjones.com>