Include all transitive swiftmodules in `bc` output group when needed
Top-level swift targets, such as `swift_test` don’t set a value for `SwiftInfo.direct_modules[0].swift`. This makes sense, as downstream targets shouldn’t get any transitive swiftmodules from it. We previously worked around this by manually including transitive direct swiftmodules. This worked fine until compiler plugins, which introduced a new way that transitive swiftmodules could be collected. Thankfully the information we need is in the `compilation_context`.

As part of this change, we now only add the `.swiftmodule` files to `generated`. Previously we would add `.swiftdoc` and the generated header as well. We can’t do that with the information in `compilation_context`. If the swift generated header is used as an input for another target, it will get picked up that way.

Signed-off-by: Brentley Jones <github@brentleyjones.com>