Depend on `--experimental_remote_download_regex` for Index Build input collection in incremental generation mode
This is a pretty big shift away from how we used to do it.

With the baggage of BwX mode gone we don’t have to track generated files in the same way. After removing that tracking we can depend on building the product for each target (i.e. the `bp` output group) to transitive generate all files needed to compile that target.

In a BwtB world though we need to make sure we still download all of those inputs, which is where `--experimental_remote_download_regex` comes in. I’ve added all of the file types that I currently know of to the regex, but it’s brittle and may require more from the user. The flag supports multiple uses, so a user can fix this on their own as well.

There are four main benefits with this new way (which is good, because the above is a pretty big downside):

1. We now generate/download previously tricky to track files like vfs overlays and hmaps
2. Less Starlark CPU usage collecting all of the generated files
3. Less Starlark memory usage retaining these generated files
4. Much smaller BEP, since the output groups are smaller, resulting in less memory usage and faster builds

Signed-off-by: Brentley Jones <github@brentleyjones.com>