Fix tree artifact `File` handling in incremental generation
We fix our handling by using a null-character terminator instead of encoding the number of paths. This is because at analysis time we can’t rely on the number of elements in a `depset`, as they might expand to more elements at execution if any of the elements are tree artifacts.

Added an example by Wiley that exercises the previous failure case.

Co-authored-by: Wiley Kestner <wiley@aya.yale.edu>
Signed-off-by: Brentley Jones <github@brentleyjones.com>