Implement support for new swift_proto_library rule (#2832)
The rules_swift PR with the new swift_proto_library rule just merged:

[https://github.com/bazelbuild/rules_swift/pull/1140](https://github.com/bazelbuild/rules_swift/pull/1140)

The provider for the new rules is backwards compatible, but now users of
the rule can define their own module names,
or if they don't, the name is derived from the swift proto library
target, not the proto library target, because multiple proto libraries
can now be compiled into a single swift proto library.

rules_xcodeproj has logic to attempt to derive this module name which
will be wrong for the new rules. This PR adds a couple of lines to read
the module name from the SwiftProtoInfo if it was set. The new rules
would actually have worked without this if the module_name attribute was
set because of the earlier code intended for swift_library targets, but
this ensures that rules_xcodeproj will still get the correct module name
even if the user doesn't customize this.

---------

Signed-off-by: Logan Shire <logan.shire@gmail.com>
Co-authored-by: Logan Shire <3255391+AttilaTheFun@users.noreply.github.com>